Fix MaxCurrentYearValidator to compare using long, not int
Number#intValue() narrows long inputs and can silently overflow,
letting out-of-range values pass validation. Compare via longValue()
instead, which safely covers short/int/long. Add explicit int/short/
long coverage plus a regression test for the overflow case.

+++ src/main/java/fi/ishtech/common/validation/validator/MaxCurrentYearValidator.java
@@ -18,6 +18,6 @@ public class MaxCurrentYearValidator implements ConstraintValidator<MaxCurrentYe
 
 	@Override
 	public boolean isValid(Number value, ConstraintValidatorContext context) {
-		return value == null || value.intValue() <= LocalDate.now().getYear();
+		return value == null || value.longValue() <= LocalDate.now().getYear();
 	}
 }
